fix(checkout): guest checkout also redirects to confirmation page
