--- client/components/NoUserCheckout.js
@@ -1,18 +1,0 @@
-import React, {Component} from 'react'
-import {Login, GuestCheckout} from './index'
-
-export class NoUserCheckout extends Component {
-  render() {
-    console.log(this.props.isLoggedIn)
-    return (
-      <div>
-        <div>Checkout as guest</div>
-        <GuestCheckout />
-        <div>Returning customers</div>
-        <Login />
-      </div>
-    )
-  }
-}
-
-export default NoUserCheckout
