+++ client/components/index.js
@@ -13,5 +13,5 @@ export {default as Signup} from './auth/SignUpForm'
 export {default as ConfirmCheckout} from './ConfirmCheckout'
 export {default as NotFound} from './error/NotFound'
 export {default as Error500} from './error/Error500'
-export {default as NoUserCheckout} from './NoUserCheckout'
 export {default as GuestCheckout} from './GuestCheckout'
+export {default as GuestCheckoutForm} from './GuestCheckoutForm'
